Move Machine files to a separate folder
