+++ src/components/Machine/index.js
@@ -1,17 +1,16 @@
 import { useEffect, useState } from "react";
 import { Container, Row, Col } from "react-bootstrap";
 
-import * as constants from "../constants.js";
-
-import Stamper from "./Stamper";
-import BiscuitLifecycle from "./biscuit/biscuitLifecycle.js";
-import BoostSwitch from "./boostSwitch.js";
-import BiscuitContainer from "./container.js";
-import Conveyor from "./conveyor.js";
-import Extruder from "./extruder.js";
-import Motor from "./motor.js";
-import Oven from "./oven.js";
-import Switch from "./switch.js";
+import * as constants from "../../constants.js";
+import Stamper from "../Stamper";
+import BiscuitLifecycle from "../biscuit/biscuitLifecycle.js";
+import BoostSwitch from "../boostSwitch.js";
+import BiscuitContainer from "../container.js";
+import Conveyor from "../conveyor.js";
+import Extruder from "../extruder.js";
+import Motor from "../motor.js";
+import Oven from "../oven.js";
+import Switch from "../switch.js";
 
 const Machine = () => {
   const [state, setState] = useState(constants.machineState.off);
